+++ boosty_downloader/src/yaml_configuration/sample_config.py
@@ -5,6 +5,8 @@ auth:
   # Insert your own cookie and auth header values here
   cookie: ''
   auth_header: ''
+  # OAuth tokens file path (will be created automatically if OAuth is used)
+  oauth_tokens_file: 'oauth_tokens.json'
 downloading_settings:
   target_directory: ./boosty-downloads
 """